backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-multipart>=0.0.9
aiofiles>=24.0.0
httpx>=0.27.0
sse-starlette>=2.0.0
openai>=1.30.0
python-dotenv>=1.0.0
